requirements.txt
Django>=3.2
djangorestframework
django-ckeditor
django-redis
googletrans==4.0.0-rc1
gunicorn
pytest-django
